refactor: remove category getters (#886)
M2-445
