+++ packages/theme/modules/catalog/category/types.d.ts
@@ -156,12 +156,6 @@ export interface CategoryTreeInterface {
   [x: string]: unknown;
 }
 
-export interface CategoryGetters<CATEGORY> {
-  getTree: (category: CATEGORY) => CategoryTreeInterface | null;
-  getBreadcrumbs?: (category: CATEGORY) => AgnosticBreadcrumb[];
-  [getterName: string]: any;
-}
-
 /** category tree implementation */
 export interface CategoryTree extends CategoryInterface, RoutableInterface {
   available_sort_by?: Maybe<Array<Maybe<Scalars['String']>>>;
