+++ packages/theme/types/core.d.ts
@@ -703,12 +703,6 @@ export interface WishlistGetters<WISHLIST, WISHLIST_ITEM> {
   [getterName: string]: (element: any, options?: any) => unknown;
 }
 
-export interface CategoryGetters<CATEGORY> {
-  getTree: (category: CATEGORY) => AgnosticCategoryTree | null;
-  getBreadcrumbs?: (category: CATEGORY) => AgnosticBreadcrumb[];
-  [getterName: string]: any;
-}
-
 export interface UserGetters<USER> {
   getFirstName: (customer: USER) => string;
   getLastName: (customer: USER) => string;
